+++ apps/frontend/src/app/[lang]/blog/[category]/[slug]/layout.tsx
@@ -33,6 +33,9 @@ export const LayoutRoute = async (props: {
 };
 
 export async function generateStaticParams() {
+  if (process.env.SKIP_BUILD_FETCH === "true") {
+    return [];
+  }
   return getArticlesSlugAndCategoryList();
 }