+++ apps/frontend/src/app/[lang]/blog/[category]/[slug]/page.tsx
@@ -28,6 +28,9 @@ export const PostRoute = async (props: {
 };
 
 export async function generateStaticParams() {
+  if (process.env.SKIP_BUILD_FETCH === "true") {
+    return [];
+  }
   return getArticlesSlugAndCategoryList();
 }